+++ plato/draw/zdog/Arrows2D.py
@@ -1,0 +1,5 @@
+from ... import draw
+from .Polygons import Polygons
+
+class Arrows2D(draw.Arrows2D, Polygons):
+    __doc__ = draw.Arrows2D.__doc__